feat(lAnalytics): LearningAnalytics as own component (support interchangeability)
