+++ libs/feature/learning-analytics/src/index.ts
@@ -1,3 +1,4 @@
 // export * from "./lib/learning-analytics";
 export * from "./lib/metrics/metrics";
 export * from "./lib/learning-analytics";
+export * from "./lib/components/unary-view";
